Consolidate test utilities into src/utils/testUtils
Move renderWithAlerts from testAlertsUtils and renderWithRouter from
PaginationLayout.test into a shared testUtils module. Update test files
to import from the new location and wrap DualListLayout tests with
router-aware rendering.
